Refactor: Separating Header's layout from Provider Information
