+++ src/components/Header/component/Provider/index.jsx
@@ -1,109 +1,49 @@
 // @flow
 import * as React from 'react'
-import { withStyles } from '@material-ui/core/styles'
-import IconButton from '@material-ui/core/IconButton'
-import ExpandLess from '@material-ui/icons/ExpandLess'
-import ExpandMore from '@material-ui/icons/ExpandMore'
-import Paragraph from '~/components/layout/Paragraph'
-import Col from '~/components/layout/Col'
-import Img from '~/components/layout/Img'
-import { type Open } from '~/components/hoc/OpenHoc'
-import { sm, md } from '~/theme/variables'
-import Identicon from '~/components/Identicon'
-
-const connectedLogo = require('../../assets/connected.svg')
+import openHoc, { type Open } from '~/components/hoc/OpenHoc'
+import ClickAwayListener from '@material-ui/core/ClickAwayListener'
+import Grow from '@material-ui/core/Grow'
+import Popper from '@material-ui/core/Popper'
+import Connected from './Connected'
+import UserDetails from './UserDetails'
 
 type Props = Open & {
   provider: string,
   network: string,
-  classes: Object,
   userAddress: string,
   connected: boolean,
-  children: Function,
 }
 
-const styles = () => ({
-  root: {
-    height: '100%',
-    display: 'flex',
-    alignItems: 'center',
-  },
-  provider: {
-    padding: `${sm} ${md}`,
-    alignItems: 'center',
-    flex: '0 1 auto',
-    display: 'flex',
-    cursor: 'pointer',
-  },
-  network: {
-    fontFamily: 'Montserrat, sans-serif',
-  },
-  logo: {
-    top: '10px',
-    position: 'relative',
-    right: '13px',
-  },
-  account: {
-    paddingRight: sm,
-    display: 'flex',
-    flexDirection: 'column',
-    justifyContent: 'center',
-  },
-  address: {
-    letterSpacing: '-0.5px',
-  },
-  expand: {
-    width: '30px',
-    height: '30px',
-  },
-})
+const Provider = ({
+  open, toggle, provider, network, userAddress, connected,
+}: Props) => (
+  <Connected
+    provider={provider}
+    network={network}
+    userAddress={userAddress}
+    connected={connected}
+    open={open}
+    toggle={toggle}
+  >
+    {providerRef => (
+      <Popper open={open} anchorEl={providerRef.current} placement="bottom-end">
+        {({ TransitionProps }) => (
+          <Grow
+            {...TransitionProps}
+          >
+            <ClickAwayListener onClickAway={toggle}>
+              <UserDetails
+                provider={provider}
+                network={network}
+                userAddress={userAddress}
+                connected={connected}
+              />
+            </ClickAwayListener>
+          </Grow>
+        )}
+      </Popper>
+    )}
+  </Connected>
+)
 
-type ProviderRef = { current: null | HTMLDivElement }
-
-class Provider extends React.Component<Props> {
-  constructor(props: Props) {
-    super(props)
-
-    this.myRef = React.createRef()
-  }
-
-  myRef: ProviderRef
-
-  render() {
-    const {
-      open, toggle, provider, network, userAddress, connected, children, classes,
-    } = this.props
-
-    const providerText = connected ? `${provider} [${network}]` : 'Not connected'
-    const cutAddress = connected ? `${userAddress.substring(0, 8)}...${userAddress.substring(36)}` : ''
-
-    return (
-      <React.Fragment>
-        <div ref={this.myRef} className={classes.root}>
-          <Col end="sm" middle="xs" className={classes.provider}>
-            { connected &&
-              <React.Fragment>
-                <Identicon address={userAddress} diameter={30} />
-                <Img className={classes.logo} src={connectedLogo} height={20} alt="Status connected" />
-              </React.Fragment>
-            }
-            <Col end="sm" middle="xs" layout="column" className={classes.account}>
-              <Paragraph size="sm" transform="capitalize" className={classes.network} noMargin bold>{providerText}</Paragraph>
-              <Paragraph size="sm" className={classes.address} noMargin>{cutAddress}</Paragraph>
-            </Col>
-            <IconButton
-              onClick={toggle}
-              disableRipple
-              className={classes.expand}
-            >
-              { open ? <ExpandLess /> : <ExpandMore />}
-            </IconButton>
-          </Col>
-        </div>
-        { children(this.myRef) }
-      </React.Fragment>
-    )
-  }
-}
-
-export default withStyles(styles)(Provider)
+export default openHoc(Provider)
