+++ src/components/Header/index.jsx
@@ -1,6 +1,8 @@
 // @flow
 import * as React from 'react'
 import { connect } from 'react-redux'
+import { logComponentStack, type Info } from '~/utils/logBoundaries'
+import Provider from './component/Provider'
 import Layout from './component/Layout'
 import actions from './actions'
 import selector from './selector'
@@ -13,9 +15,23 @@ type Props = {
   connected: boolean,
 }
 
-class Header extends React.PureComponent<Props> {
+type State = {
+  hasError: boolean,
+}
+
+class Header extends React.PureComponent<Props, State> {
+  state = {
+    hasError: false,
+  }
+
   componentDidMount() {
     this.props.fetchProvider()
+  }
+
+  componentDidCatch(error: Error, info: Info) {
+    this.setState({ hasError: true })
+
+    logComponentStack(error, info)
   }
 
   reloadWallet = () => {
@@ -27,14 +43,26 @@ class Header extends React.PureComponent<Props> {
       provider, userAddress, network, connected,
     } = this.props
 
+    const { hasError } = this.state
+
+    // const providerDisconnected = !hasError && !connected
+    const providerConnected = !hasError && connected
+
     return (
       <Layout
-        provider={provider}
         reloadWallet={this.reloadWallet}
-        userAddress={userAddress}
-        network={network}
-        connected={connected}
-      />
+      >
+        {/* hasError && <ProviderError /> */}
+        {/* providerDisconnected && <ProviderDisconnected /> */}
+        { providerConnected &&
+          <Provider
+            provider={provider}
+            userAddress={userAddress}
+            network={network}
+            connected={connected}
+          />
+        }
+      </Layout>
     )
   }
 }
